updated api + some fixes

See merge request students-giftery/students-giftery!31

+++ src/main/java/ru/itis/studentsgiftery/services/impl/BalanceServiceImpl.java
@@ -23,7 +23,7 @@ public class BalanceServiceImpl implements BalanceService {
     private final AccountMapper accountMapper;
 
     @Override
-    public Integer getAccountBalance(Long accountId) {
+    public Long getAccountBalance(Long accountId) {
         Account account = accountsRepository.findById(accountId).orElseThrow((Supplier<RuntimeException>) ()
                 -> new AccountNotFoundException("Account not found")
         );
